fix: request permission in SaveAsActivity on first-run (#271)
* Fix "Replace characters on save as function, FileNotFoundException" #250

* Update CHANGELOG.md

update changelog

* Update app/src/main/kotlin/org/fossify/filemanager/activities/SaveAsActivity.kt


* Update CHANGELOG.md


* Update CHANGELOG.md


* Merge branch 'FossifyOrg:main' into main


* Update CHANGELOG.md

* fix PR Feedback


* fix duplicated code by moving it to SimpleActivity.kt


* moved handleStoragePermissions() to SimpleActivity.kt from MainActivity.kt
introduced tryInitFileManager() to SaveAsActivity.kt
removed const MANAGE_STORAGE_RC from MainActivity.kt


* adjusted toast in try catch in storage permission action


* catch ActivityNotFoundException with fallback of SecurityException


* Apply suggestions from code review

thanks for the feedback!


* remove indent


* adjust changelog


---------


Refs: https://github.com/FossifyOrg/File-Manager/issues/85
