+++ app/src/main/kotlin/org/fossify/filemanager/activities/SaveAsActivity.kt
@@ -18,7 +18,21 @@ class SaveAsActivity : SimpleActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(binding.root)
+        tryInitFileManager()
+    }
 
+    private fun tryInitFileManager() {
+        handleStoragePermission { granted ->
+            if (granted) {
+                saveAsDialog()
+            } else {
+                toast(R.string.no_storage_permissions)
+                finish()
+            }
+        }
+    }
+
+    private fun saveAsDialog() {
         if (intent.action == Intent.ACTION_SEND && intent.extras?.containsKey(Intent.EXTRA_STREAM) == true) {
             FilePickerDialog(this, pickFile = false, showHidden = config.shouldShowHidden(), showFAB = true, showFavoritesButton = true) {
                 val destination = it
