+++ app/src/main/kotlin/org/fossify/filemanager/activities/SimpleActivity.kt
@@ -1,9 +1,15 @@
 package org.fossify.filemanager.activities
 
 import android.annotation.SuppressLint
+import android.content.Intent
 import android.os.Environment
+import android.provider.Settings
 import org.fossify.commons.activities.BaseSimpleActivity
+import org.fossify.commons.dialogs.ConfirmationAdvancedDialog
 import org.fossify.commons.extensions.hasPermission
+import android.net.Uri
+import org.fossify.commons.extensions.showErrorToast
+import org.fossify.commons.extensions.toast
 import org.fossify.commons.helpers.PERMISSION_WRITE_STORAGE
 import org.fossify.commons.helpers.isRPlus
 import org.fossify.filemanager.R
@@ -31,9 +37,22 @@ open class SimpleActivity : BaseSimpleActivity() {
         R.mipmap.ic_launcher_grey_black
     )
 
+    companion object {
+        private const val MANAGE_STORAGE_RC = 201
+    }
+
     override fun getAppLauncherName() = getString(R.string.app_launcher_name)
 
     override fun getRepositoryName() = "File-Manager"
+
+   @SuppressLint("NewApi")
+    override fun onActivityResult(requestCode: Int, resultCode: Int, resultData: Intent?) {
+        super.onActivityResult(requestCode, resultCode, resultData)
+        isAskingPermissions = false
+        if (requestCode == MANAGE_STORAGE_RC && isRPlus()) {
+            actionOnPermission?.invoke(Environment.isExternalStorageManager())
+        }
+    }
 
     @SuppressLint("NewApi")
     fun hasStoragePermission(): Boolean {
@@ -41,6 +60,41 @@ open class SimpleActivity : BaseSimpleActivity() {
             Environment.isExternalStorageManager()
         } else {
             hasPermission(PERMISSION_WRITE_STORAGE)
+        }
+    }
+
+    @SuppressLint("InlinedApi")
+    fun handleStoragePermission(callback: (granted: Boolean) -> Unit) {
+        actionOnPermission = null
+        if (hasStoragePermission()) {
+            callback(true)
+        } else {
+            if (isRPlus()) {
+                ConfirmationAdvancedDialog(this, "", R.string.access_storage_prompt, R.string.ok, 0, false) { success ->
+                    if (success) {
+                        isAskingPermissions = true
+                        actionOnPermission = callback
+                        try {
+                            val intent = Intent(Settings.ACTION_MANAGE_APP_ALL_FILES_ACCESS_PERMISSION)
+                            intent.addCategory("android.intent.category.DEFAULT")
+                            intent.data = Uri.parse("package:$packageName")
+                            startActivityForResult(intent, MANAGE_STORAGE_RC)
+                        } catch (e: android.content.ActivityNotFoundException) {
+                            showErrorToast(e)
+                            val intent = Intent()
+                            intent.action = Settings.ACTION_MANAGE_ALL_FILES_ACCESS_PERMISSION
+                            startActivityForResult(intent, MANAGE_STORAGE_RC)
+                        } catch (e: SecurityException) {
+                            showErrorToast(e)
+                            finish()
+                        }
+                    } else {
+                        finish()
+                    }
+                }
+            } else {
+                handlePermission(PERMISSION_WRITE_STORAGE, callback)
+            }
         }
     }
 }
